(#3176) - increment last_seq only for active changes
Matching CouchDB behaviour, the last_seq field in _changes should
represent the last active change processed.

The previous implementation was correct for the WebSQL adapter. However,
the IndexedDB and LevelDB adapters would update last_seq for each change
regardless of whether there was a more recent version of the document
in the feed (which would subsequently be filtered out).
This worked for ascending changes but when querying the feed in descending
order, the bug was exposed.